(refactor) Comment out deprecated files, update naming of relevant ones in viewWorkout
