+++ client/scripts/components/viewWorkout/DEPR_viewInstructions.js
@@ -1,0 +1,66 @@
+//DEPR as of 1/19/16
+
+// /*
+// * @Author: vincetam
+// * @Date:   2015-12-29 14:32:54
+// * @Last Modified by:   VINCE
+// * @Last Modified time: 2016-01-19 09:18:23
+// */
+
+// 'use strict';
+
+// var React = require('react-native');
+// var editWorkoutActions = require('../../actions/editWorkoutActions');
+
+// var {
+//   StyleSheet,
+//   View,
+//   TextInput,
+// } = React;
+
+// var ViewInstructions = React.createClass({
+//   setInstructions: function(instructions) {
+//     editWorkoutActions.setInstructions(instructions, this.props.partIdx);
+//   },
+
+//   render: function(){
+//     return (
+//       /* jshint ignore:start */
+//       <View style={styles.instructionsContainer}>
+//         <TextInput
+//           value={this.props.instructions}
+//           placeholder='Instructions'
+//           onChangeText={(text) => this.setInstructions(text)}
+//           multiline={true}
+//           style={styles.instructionsTextInput} />
+//         <View style={styles.cellSeparatorBackground}>
+//           <View style={styles.cellSeparatorLine}></View>
+//         </View>
+//       </View>
+//       /* jshint ignore:end */
+//     );
+//   }
+// });
+
+// var styles = StyleSheet.create({
+//   instructionsContainer: {
+//     marginTop: 10,
+//     marginLeft: 10,
+//   },
+//   instructionsTextInput: {
+//     height: 40,
+//     fontFamily: 'Avenir Next',
+//     color: '#2D2D2D',
+//     fontStyle: 'italic',
+//     fontSize: 16
+//   },
+//   cellSeparatorBackground: {
+//     backgroundColor: '#fff',
+//   },
+//   cellSeparatorLine: {
+//     height: 0.5,
+//     backgroundColor: '#c8c7cc',
+//   }
+// });
+
+// module.exports = ViewInstructions;
