+++ client/scripts/components/viewWorkout/exerciseView.js
