+++ client/scripts/components/viewWorkout/instructionsView.js
