+++ client/scripts/components/viewWorkout/workoutTab.js
@@ -2,14 +2,10 @@
 
 var React = require('react-native');
 var Subscribable = require('Subscribable'); //used for addListenerOn
-var editWorkoutActions = require('../../actions/editWorkoutActions');
-var viewWorkoutActions = require('../../actions/viewWorkoutActions');
 var modalActions = require('../../actions/modalActions');
 
 //Load components
-var WorkoutChoice = require('./_workoutChoice');
-var ViewWorkout = require('./viewWorkout');
-var EditWorkout = require('../editWorkout/editWorkout');
+var WorkoutChoice = require('./workoutChoice');
 
 var {
   StyleSheet,
@@ -71,48 +67,23 @@ var WorkoutTab = React.createClass({
 });
 
 var NavBarRouteMapper = {
-  LeftButton: function(route, navigator, index, navState) {
-    //Show left button based on user's view
-    switch (route.name) {
-      //If viewing today's workout, render create workout button
-      case 'New Workout':
-        var handleLeftPress = function(){
-          viewWorkoutActions.initPartsAreLogged();
+   LeftButton: function(route, navigator, index, navState) {
+    return (
+      /* jshint ignore:start */
+      <TouchableOpacity
+        style={styles.navBarComponentContainer}
+        onPress={ () => {
           if(index > 0) {
             navigator.pop();
           }
-        };
-        return (
-          /* jshint ignore:start */
-          <TouchableOpacity
-            style={styles.navBarComponentContainer}
-            onPress={ () => handleLeftPress()}>
-            { index > 0 ?
-              <Image
-                style={{height: 22, width: 12}}
-                source={ require('image!backArrow') } /> : null }
-          </TouchableOpacity>
-          /* jshint ignore:end */
-        );
-        break;
-      default:
-        return (
-          /* jshint ignore:start */
-          <TouchableOpacity
-            style={styles.navBarComponentContainer}
-            onPress={ () => {
-              if(index > 0) {
-                navigator.pop();
-              }
-            }}>
-            { index > 0 ?
-              <Image
-                style={{height: 22, width: 12}}
-                source={ require('image!backArrow') } /> : null }
-          </TouchableOpacity>
-          /* jshint ignore:end */
-        );
-    }
+        }}>
+        { index > 0 ?
+          <Image
+            style={{height: 22, width: 12}}
+            source={ require('image!backArrow') } /> : null }
+      </TouchableOpacity>
+      /* jshint ignore:end */
+    );
   },
 
   RightButton: function(route, navigator, index, navState) {
@@ -128,16 +99,6 @@ var NavBarRouteMapper = {
             onPress={() => handleRightPress()}
             style={styles.navBarComponentContainer} >
             <Image source={require('image!createIcon')} />
-          </TouchableOpacity>
-        );
-        break;
-      //If in New Workout scene, render 'Add Part' button
-      case 'New Workout':
-        return (
-          <TouchableOpacity
-            onPress={() => editWorkoutActions.addPart()}
-            style={styles.navBarComponentContainer} >
-            <Text style={styles.navBarSideText}>Add Part</Text>
           </TouchableOpacity>
         );
         break;
